adding history data types: 1. Historical Volatility 2. Option Implied Volatility 3. Trades
fix tests
upgrade schema
disable old schema upgrade tests

+++ db/src/main/scala/serv1/db/schema/TickerData.scala
@@ -5,7 +5,7 @@ import serv1.model.HistoricalData
 
 import scala.language.implicitConversions
 
-case class TickerData(id: Int, time: Long, open: Long, high: Long, low: Long, close: Long, vol: Double, historicalDataType: HistoricalDataType)
+case class TickerData(id: Int, time: Long, open: Double, high: Double, low: Double, close: Double, vol: Double, historicalDataType: HistoricalDataType)
 
 object TickerData {
   implicit def historicalDataToTickerData(historicalData: HistoricalData): TickerData = {
